refactor: re-home GitLab CI TestOps adapter to providers/gitlab (CHAOS-1556)
